requirements.txt
alembic==1.2.1
apache-airflow==1.10.5
apispec==2.0.2
attrs==19.2.0
Babel==2.7.0
boto3==1.9.247
botocore==1.12.247
cached-property==1.5.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
colorama==0.4.1
colorlog==4.0.2
configparser==3.5.3
contextlib2==0.6.0.post1
croniter==0.3.30
dill==0.2.9
docutils==0.15.2
dumb-init==1.2.2
enum34==1.1.6
Flask==1.1.1
Flask-Admin==1.5.3
Flask-AppBuilder==1.13.1
Flask-Babel==0.12.2
Flask-Caching==1.3.3
Flask-JWT-Extended==3.24.0
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.4.1
flask-swagger==0.2.13
Flask-WTF==0.14.2
funcsigs==1.0.0
functools32==3.2.3.post2
future==0.16.0
futures==3.3.0
gunicorn==19.9.0
idna==2.8
importlib-metadata==0.23
iso8601==0.1.12
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
json-merge-patch==0.2
jsonschema==3.1.1
lazy-object-proxy==1.4.2
lockfile==0.12.2
Mako==1.1.0
Markdown==2.6.11
MarkupSafe==1.1.1
marshmallow==2.19.5
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.17.1
monotonic==1.5
more-itertools==5.0.0
numpy==1.16.5
ordereddict==1.1
pandas==0.24.2
pathlib2==2.3.5
pendulum==1.4.4
prison==0.1.0
psutil==5.6.3
psycopg2==2.8.3
py4j==0.10.7
Pygments==2.4.2
PyJWT==1.7.1
pyrsistent==0.15.4
pyspark==2.4.4
python-daemon==2.1.2
python-dateutil==2.8.0
python-editor==1.0.4
python-openid==2.2.5
pytz==2019.3
pytzdata==2019.3
PyYAML==5.1.2
requests==2.22.0
s3transfer==0.2.1
scandir==1.10.0
setproctitle==1.1.10
six==1.12.0
spark==0.2.1
SQLAlchemy==1.3.10
tabulate==0.8.5
tenacity==4.12.0
termcolor==1.1.0
text-unidecode==1.2
thrift==0.11.0
typing==3.7.4.1
tzlocal==1.5.1
unicodecsv==0.14.1
urllib3==1.25.6
Werkzeug==0.16.0
WTForms==2.2.1
zipp==0.6.0
zope.deprecation==4.4.0
